backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
django==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
celery==5.3.6
redis==5.0.1
openai==1.10.0
anthropic==0.18.1
beautifulsoup4==4.12.3
playwright==1.41.2
requests==2.31.0
python-dotenv==1.0.0
pillow==10.2.0
pandas==2.2.0
numpy==1.26.3

